Few modifications to get ATA read of kernel into memory working
issue was that we were booting from a floppy and not a harddisk as floppies are not controlled by the ATA bus. Modified bochs to use the os.bin file as a harddrive and everything works as it should now. Will go back and revert some of the ATA driver changes that were undone for troubleshooting.