Update make structure
Updated the makefile to be more robust and future proof and to be a lot more hands off moving forward (made it smart! :)). Updated the linker script to always load place the .text._start section at the beginning of the output binary(fixes issue with kernel.asm _start function not being properly placed at 0x100000 and the entire kernel shitting itself after the bootloader hand off control). Updated the kernel.asm file with this new section so the linker can identify it.