GDB was loosing its shit when trying to debug, I think it has been fixed now
To fix this I had to undo the linker trick I was using to ensure that the kernel.asm.o code was being loaded first thing in the final binary. Instead I had to update the make file to remove kernel.asm.o from the object file list and then re-add it to the front to ensure it is always the first argument to be processed by the linker, which in turn makes it the first object of the final combined object file. I have no clue why this breaks GDB but after these changes GDB works much better(still a few weird issues with next behaving as step). The issue with GDB was that it was not properly stopping at break points inside functions. It would stop at the entry of a function, but not stop at c source code lines inside the function.